requirements.txt
openai
cryptography
pyyaml
psycopg2-binary